Extract text grabbing part of getCompletions to readApiCall.ts
